Get information about a single Facebook video directly from the API. Mostly used internally.

See the official documentation for reference: https://developers.facebook.com/docs/graph-api/reference/video/video_insights/

Usage

cc_api_get_fb_video_insights(
  fb_video_id,
  metrics = cc_valid_fields_fb_video_insights,
  cache = TRUE,
  update = TRUE,
  api_version = "v18.0",
  fb_page_id = NULL,
  fb_page_token = NULL
)

Arguments

fb_video_id

Instagram media identifier, must be a vector of length 1. A list of identifiers for your account can be retrieved with `cc_get_fb_page_video()`.

metrics

Metrics to be retrieved. Consider that depending on the media type, different media types are effectively available. Requesting the wrong metrics will cause an error. Defaults to NULL. If left to NULL, metrics will be chosen based on the media type. See the official documentation for reference: https://developers.facebook.com/docs/graph-api/reference/insights/#page-posts

fb_page_token